+++ src/app.d.ts
@@ -14,6 +14,7 @@ declare global {
 		type Auth = import('$lib/server/lucia').Auth;
 		type DatabaseUserAttributes = {
 			username: string;
+			email: string|null
 		};
 		type DatabaseSessionAttributes = {};
 	}